refactor bridge lock
