+++ modules/cli/cmd/sequencer/helpers_test.go
@@ -35,3 +35,43 @@ func TestAddPortToURL(t *testing.T) {
 		assert.Equal(t, want, got)
 	})
 }
+
+func TestConvertToUint128(t *testing.T) {
+	testCases := []struct {
+		name        string
+		input       string
+		expectedLo  uint64
+		expectedHi  uint64
+		expectError bool
+	}{
+		{"Zero", "0", 0, 0, false},
+		{"Boundary64", "18446744073709551615", 18446744073709551615, 0, false}, // 2^64 - 1
+		// {lo: 0, hi: 1} = 0b00000000000000000000000000000000000000000000000000000000000000001 (1 in 65th) = 0d18446744073709551616
+		{"Boundary64PlusOne", "18446744073709551616", 0, 1, false}, // 2^64
+		// 0d18446744073709551615 = 0b1111111111111111111111111111111111111111111111111111111111111111 (64 bits),
+		{"Boundary128", "340282366920938463463374607431768211455", 18446744073709551615, 18446744073709551615, false}, // 2^128 - 1
+		// this number is too big to handle!
+		{"Boundary128PlusOne", "340282366920938463463374607431768211456", 0, 0, true}, // 2^128
+		{"Small", "1234", 1234, 0, false},
+		{"NonNumeric", "nonnumeric", 0, 0, true},
+		{"Negative", "-1", 0, 0, true},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			uint128, err := convertToUint128(tc.input)
+			if tc.expectError {
+				if err == nil {
+					t.Errorf("Expected an error but did not get one")
+				}
+			} else {
+				if err != nil {
+					t.Errorf("Did not expect an error but got one: %v", err)
+				}
+				if uint128.Lo != tc.expectedLo || uint128.Hi != tc.expectedHi {
+					t.Errorf("Expected (%d, %d), got (%d, %d)", tc.expectedLo, tc.expectedHi, uint128.Lo, uint128.Hi)
+				}
+			}
+		})
+	}
+}
